feat: 바텀sheetDialog 연결, 이후 바텀sheetDialog에서 카메라 작동되게 변경
