+++ app/src/main/java/com/konkuk/nongboohae/presentation/main/MainActivity.kt
@@ -11,9 +11,11 @@ import android.util.Log
 import androidx.core.app.ActivityCompat
 import androidx.core.content.ContextCompat
 import androidx.core.content.FileProvider
+import androidx.fragment.app.DialogFragment
 import com.konkuk.nongboohae.R
 import com.konkuk.nongboohae.databinding.ActivityMainBinding
 import com.konkuk.nongboohae.presentation.base.BaseActivity
+import com.konkuk.nongboohae.presentation.diagnosis.DiagnosisBottomSheet
 import java.io.File
 import java.io.IOException
 import java.text.SimpleDateFormat
@@ -23,10 +25,6 @@ import java.util.Date
 class MainActivity : BaseActivity<ActivityMainBinding>() {
     override val TAG: String = "MainActivity"
     override val layoutRes: Int = R.layout.activity_main
-    private val CAMERA_PERMISSION_REQUEST_CODE = 100
-    val REQUEST_TAKE_PHOTO = 1
-    lateinit var currentPhotoPath: String
-    private var getPhotoURI: Uri? = null
 
     override fun initViewModel() {
 
@@ -34,75 +32,9 @@ class MainActivity : BaseActivity<ActivityMainBinding>() {
 
     override fun afterViewCreated() {
         binding.fab.setOnClickListener {
-            requestCameraPermission()
+            val modal = DiagnosisBottomSheet()
+            modal.setStyle(DialogFragment.STYLE_NORMAL, R.style.TransParentBottomSheetDialogTheme)
+            modal.show(supportFragmentManager, DiagnosisBottomSheet.TAG)
         }
     }
-
-    // 카메라 권한 요청
-    private fun requestCameraPermission() {
-        if (ContextCompat.checkSelfPermission(this, android.Manifest.permission.CAMERA) != PackageManager.PERMISSION_GRANTED) {
-            ActivityCompat.requestPermissions(this, arrayOf(android.Manifest.permission.CAMERA), CAMERA_PERMISSION_REQUEST_CODE)
-        } else {
-            callCamera() // 이미 권한이 허용된 경우
-        }
-    }
-
-    // 권한 요청 결과 처리
-    override fun onRequestPermissionsResult(requestCode: Int, permissions: Array<out String>, grantResults: IntArray) {
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults)
-        if (requestCode == CAMERA_PERMISSION_REQUEST_CODE) {
-            if (grantResults.isNotEmpty() && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-                callCamera()
-            } else {
-                showToast("카메라 권한을 승인해 주세요")
-            }
-        }
-    }
-
-    private fun callCamera() {
-        Intent(MediaStore.ACTION_IMAGE_CAPTURE).also { takePictureIntent ->
-            takePictureIntent.resolveActivity(packageManager)?.also {
-                val photoFile: File? = try {
-                    createImageFile()
-                } catch (ex: IOException) {
-                    null
-                }
-                // Continue only if the File was successfully created
-                photoFile?.also {
-                    val photoURI: Uri = FileProvider.getUriForFile(
-                        this,
-                        "com.konkuk.nongboohae.fileprovider",
-                        it
-                    )
-                    getPhotoURI = photoURI
-                    takePictureIntent.putExtra(MediaStore.EXTRA_OUTPUT, photoURI)
-                    startActivityForResult(takePictureIntent, REQUEST_TAKE_PHOTO)
-                }
-            }
-        }
-    }
-
-    @Throws(IOException::class)
-    private fun createImageFile(): File {
-        // Create an image file name
-        val timeStamp: String = SimpleDateFormat("yyyyMMdd_HHmmss").format(Date())
-        val storageDir: File = getExternalFilesDir(Environment.DIRECTORY_PICTURES)!!
-        return File.createTempFile(
-            "JPEG_${timeStamp}_", /* prefix */
-            ".jpg", /* suffix */
-            storageDir /* directory */
-        ).apply {
-            // Save a file: path for use with ACTION_VIEW intents
-            currentPhotoPath = absolutePath
-        }
-    }
-
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        super.onActivityResult(requestCode, resultCode, data)
-        if (requestCode == REQUEST_TAKE_PHOTO && resultCode == RESULT_OK) {
-            Log.d("ImageURI", getPhotoURI.toString())
-            // 이미지 URI 획득!!
-        }
-    }
-
 }
